+++ org.hl7.fhir.r5/src/main/java/org/hl7/fhir/r5/conformance/ProfileUtilities.java
@@ -410,6 +410,13 @@ public class ProfileUtilities extends TranslatingUtilities {
 
   public interface ProfileKnowledgeProvider {
     class BindingResolution {
+      public BindingResolution(String display, String url) {
+        this.display = display;
+        this.url = url;
+      }
+      public BindingResolution() {
+        // TODO Auto-generated constructor stub
+      }
       public String display;
       public String url;
     }
@@ -750,14 +757,16 @@ public class ProfileUtilities extends TranslatingUtilities {
         }
         if (!Utilities.noString(b.toString())) {
           String msg = "The profile "+derived.getUrl()+" has "+ce+" "+Utilities.pluralize("element", ce)+" in the differential ("+b.toString()+") that don't have a matching element in the snapshot: check that the path and definitions are legal in the differential (including order)";
-          System.out.println("Error in snapshot generation: "+msg);
-          if (!debug) {
-            System.out.println("Differential: ");
-            for (ElementDefinition ed : derived.getDifferential().getElement())
-              System.out.println("  "+ed.getId()+" = "+ed.getPath()+" : "+typeSummaryWithProfile(ed)+"["+ed.getMin()+".."+ed.getMax()+"]"+sliceSummary(ed)+"  "+constraintSummary(ed));
-            System.out.println("Snapshot: ");
-            for (ElementDefinition ed : derived.getSnapshot().getElement())
-              System.out.println("  "+ed.getId()+" = "+ed.getPath()+" : "+typeSummaryWithProfile(ed)+"["+ed.getMin()+".."+ed.getMax()+"]"+sliceSummary(ed)+"  "+constraintSummary(ed));
+          if (debug) {
+            System.out.println("Error in snapshot generation: "+msg);
+            if (!debug) {
+              System.out.println("Differential: ");
+              for (ElementDefinition ed : derived.getDifferential().getElement())
+                System.out.println("  "+ed.getId()+" = "+ed.getPath()+" : "+typeSummaryWithProfile(ed)+"["+ed.getMin()+".."+ed.getMax()+"]"+sliceSummary(ed)+"  "+constraintSummary(ed));
+              System.out.println("Snapshot: ");
+              for (ElementDefinition ed : derived.getSnapshot().getElement())
+                System.out.println("  "+ed.getId()+" = "+ed.getPath()+" : "+typeSummaryWithProfile(ed)+"["+ed.getMin()+".."+ed.getMax()+"]"+sliceSummary(ed)+"  "+constraintSummary(ed));
+            }
           }
           if (exception)
             throw new DefinitionException(msg);
@@ -2528,8 +2537,12 @@ public class ProfileUtilities extends TranslatingUtilities {
           generateSnapshot(context.fetchTypeDefinition("Extension"), sd, sd.getUrl(), webUrl, sd.getName());
         }
       }
-      if (sd == null)
-        System.out.println("Failed to find referenced profile: " + type.getProfile());
+      if (sd == null) {
+        if (debug) {
+          System.out.println("Failed to find referenced profile: " + type.getProfile());
+        }
+      }
+        
     }
     if (sd == null)
       sd = context.fetchTypeDefinition(type.getWorkingCode());
@@ -5244,8 +5257,8 @@ public class ProfileUtilities extends TranslatingUtilities {
           for (ElementDefinitionConstraintComponent inv : definition.getConstraint()) {
             if (!c.getPieces().isEmpty()) { c.addPiece(gen.new Piece("br")); }
             c.getPieces().add(checkForNoChange(inv, gen.new Piece(null, inv.getKey()+": ", null).addStyle("font-weight:bold")));
-            if (inv.getHumanElement().hasExtension("http://hl7.org/fhir/StructureDefinition/rendering-markdown")) {
-              c.addMarkdown(inv.getHumanElement().getExtensionString("http://hl7.org/fhir/StructureDefinition/rendering-markdown"));
+            if (inv.getHumanElement().hasExtension(ToolingExtensions.EXT_REND_MD)) {
+              c.addMarkdown(inv.getHumanElement().getExtensionString(ToolingExtensions.EXT_REND_MD));
             } else {
               c.getPieces().add(checkForNoChange(inv, gen.new Piece(null, inv.getHuman(), null)));
             }
@@ -6720,8 +6733,9 @@ public class ProfileUtilities extends TranslatingUtilities {
   }
 
 
-  public void setNewSlicingProcessing(boolean newSlicingProcessing) {
+  public ProfileUtilities setNewSlicingProcessing(boolean newSlicingProcessing) {
     this.newSlicingProcessing = newSlicingProcessing;
+    return this;
   }