+++ org.hl7.fhir.r5/src/main/java/org/hl7/fhir/r5/renderers/DataRenderer.java
@@ -134,7 +134,7 @@ public class DataRenderer extends Renderer {
       }
   
       // 2. markdown
-      String s = getContext().getMarkdown().process(Utilities.escapeXml(text), "narrative generator");
+      String s = getContext().getMarkdown().process(text, "narrative generator");
       XhtmlParser p = new XhtmlParser();
       XhtmlNode m;
       try {
@@ -346,7 +346,7 @@ public class DataRenderer extends Renderer {
         "ExampleScenario", "GraphDefinition", "ImplementationGuide", "Library", "Measure", "MessageDefinition", "NamingSystem", "PlanDefinition"
         ))
       return true;
-    return sd.getBaseDefinitionElement().hasExtension("http://hl7.org/fhir/StructureDefinition/structuredefinition-codegen-super");
+    return false;
   }
 
   // -- 4. Language support ------------------------------------------------------