+++ org.hl7.fhir.r5/src/main/java/org/hl7/fhir/r5/utils/IGHelper.java
@@ -36,12 +36,12 @@ import org.hl7.fhir.r5.model.ImplementationGuide.ImplementationGuideDefinitionPa
 public class IGHelper {
 
   
-  public static final String EXT_SPREADSHEET = ToolingExtensions.EXT_IGP_SPREADSHEET;
-  public static final String EXT_MAPPING_CSV = ToolingExtensions.EXT_IGP_MAPPING_CSV;
-  public static final String EXT_BUNDLE = ToolingExtensions.EXT_IGP_BUNDLE;
-  public static final String EXT_RESOURCE_INFO = ToolingExtensions.EXT_IGP_RESOURCE_INFO;
-  public static final String EXT_CONTAINED_RESOURCE_INFO = ToolingExtensions.EXT_IGP_CONTAINED_RESOURCE_INFO;
-  public static final String EXT_PRIVATE_BASE = ToolingExtensions.EXT_PRIVATE_BASE;
+//  public static final String EXT_SPREADSHEET = ToolingExtensions.EXT_IGP_SPREADSHEET;
+//  public static final String EXT_MAPPING_CSV = ToolingExtensions.EXT_IGP_MAPPING_CSV;
+//  public static final String EXT_BUNDLE = ToolingExtensions.EXT_IGP_BUNDLE;
+//  public static final String EXT_RESOURCE_INFO = ToolingExtensions.EXT_IGP_RESOURCE_INFO;
+//  public static final String EXT_CONTAINED_RESOURCE_INFO = ToolingExtensions.EXT_IGP_CONTAINED_RESOURCE_INFO;
+//  public static final String EXT_PRIVATE_BASE = ToolingExtensions.EXT_PRIVATE_BASE;
 
   public static String readStringParameter(ImplementationGuideDefinitionComponent ig, String name) {
     for (ImplementationGuideDefinitionParameterComponent p : ig.getParameter()) {