+++ RELEASE_NOTES.md
@@ -1,7 +1,8 @@
 ## Validator Changes
 
-* no changes
+* Work around R4 issue with http://www.rfc-editor.org/bcp/bcp13.txt value set (should be http://hl7.org/fhir/ValueSet/mimetypes)
+* Minor improvements to error messages
 
 ## Other code changes
 
-* no changes
+n/a